requirements.txt
feedparser
configparser
openai
jinja2
beautifulsoup4
pyyaml
python-dotenv